refactoring the server code
