+++ bittensor/_neuron/text/core_server/__init__.py
@@ -24,9 +24,22 @@ Example:
 
 import bittensor
 import os
-
+import torch
 from .nucleus_impl import server
-from .run import serve
+from prometheus_client import Counter, Gauge, Histogram, Summary, Info, CollectorRegistry
+from threading import Lock
+from loguru import logger; logger = logger.opt(colors=True)
+import time
+
+from datetime import datetime,timedelta
+import wandb
+import pandas
+
+# Torch
+import torch
+import torch.nn.functional as F
+from torch.nn.utils.rnn import pad_sequence
+from torch.nn.utils import clip_grad_norm_
 
 class neuron:
     r"""
@@ -54,9 +67,9 @@ class neuron:
             seq2seq (:obj:bittensor.metagraph, `optional`):
                 seq2seq synapse control
             synapse_list (:obj:list of int, `optional`):
-                
+      
 
-    Examples:: 
+    Examples::
             >>> subtensor = bittensor.subtensor(network='nakamoto')
             >>> server = bittensor.neuron.text.core_server.neuron(subtensor=subtensor)
             >>> server.run()
@@ -74,7 +87,9 @@ class neuron:
         seq2seq = None,
         synapse_list = None,
     ):
-        if config == None: config = server.config()
+        if config is None:
+            config = server.config()
+
         config = config; 
 
         if synapse_list != None:
@@ -111,25 +126,35 @@ class neuron:
             config = config,
             port = config.prometheus.port if config.axon.port == bittensor.defaults.axon.port else config.axon.port - 1000
         )
+        # --- Setup prometheus summaries.
+        # These will not be posted if the user passes --prometheus.level OFF
+        registry = CollectorRegistry()
+        self.prometheus_counters = Counter('neuron_counters', 'Counter sumamries for the running server-miner.', ['neuron_counters_name'], registry=registry)
+        self.prometheus_guages = Gauge('neuron_guages', 'Guage sumamries for the running server-miner.', ['neuron_guages_name'], registry=registry)
+        self.prometheus_info = Info('neuron_info', "Info sumamries for the running server-miner.", registry=registry)
 
-        self.model = server(config = config)
+        self.mutex = Lock()
+        self.model = server(config = config).to(config.neuron.device)
         self.config = config
         self.config.to_prometheus()
 
-        self.subtensor = subtensor
-        self.wallet = wallet
-        self.axon = axon
-        self.metagraph = metagraph
+        self.subtensor = bittensor.subtensor(config = config) if subtensor == None else subtensor
+        self.wallet = bittensor.wallet( config = config ) if wallet == None else wallet
+        self.metagraph = bittensor.metagraph ( config = config, subtensor = subtensor) if metagraph == None else metagraph
 
-    def run(self):
-        serve(
-            self.config,
-            self.model,
-            subtensor = self.subtensor,
-            wallet = self.wallet,
-            axon = self.axon,
-            metagraph = self.metagraph,
-        )
+        if axon == None:
+            axon = bittensor.axon(
+                config = config,
+                wallet = wallet,
+                synapse_checks=self.synapse_check,
+                synapse_last_hidden = self.forward_hidden_state if self.model.config.neuron.lasthidden else None,
+                synapse_causal_lm = self.forward_casual_lm if self.model.config.neuron.causallm else None,
+                synapse_causal_lm_next = self.forward_casual_lm_next if self.model.config.neuron.causallmnext else None,
+                synapse_seq_2_seq = self.forward_generate if self.model.config.neuron.seq2seq else None ,
+                blacklist = self.blacklist if not self.model.config.neuron.disable_blacklist else None,
+                priority = self.priority if not self.model.config.neuron.disable_priority else None,
+            )
+        self.axon = axon
 
 
     @classmethod
@@ -152,3 +177,406 @@ class neuron:
         config.neuron.full_path = os.path.expanduser(full_path)
         if not os.path.exists(config.neuron.full_path):
             os.makedirs(config.neuron.full_path)
+
+    def run(
+            self,
+        ):
+        self.config.to_defaults()
+
+        # Load/Create our bittensor wallet.
+        self.wallet.reregister(subtensor=self.subtensor)
+
+
+        self.metagraph.load().sync().save()
+
+        # Create our optimizer.
+        optimizer = torch.optim.SGD(
+            [ {"params": self.model.parameters()} ],
+            lr = self.config.neuron.learning_rate,
+            momentum = self.config.neuron.momentum,
+        )
+
+        self.prometheus_guages.labels( 'model_size_params' ).set( sum(p.numel() for p in self.model.parameters()) )
+        self.prometheus_guages.labels( 'model_size_bytes' ).set( sum(p.element_size() * p.nelement() for p in self.model.parameters()) )
+        self.prometheus_info.info ({
+            'type': "core_server",
+            'uid': str(self.metagraph.hotkeys.index( self.wallet.hotkey.ss58_address )),
+            'network': self.config.subtensor.network,
+            'coldkey': str(self.wallet.coldkeypub.ss58_address),
+            'hotkey': str(self.wallet.hotkey.ss58_address),
+        })
+
+        self.timecheck_dicts = {bittensor.proto.RequestType.FORWARD:{}, bittensor.proto.RequestType.BACKWARD:{}}
+
+        # Create our axon server and subscribe it to the network.
+        self.axon.start().serve(subtensor=self.subtensor)
+        self.axon.attach_backward_callback(self.backward_callback)
+
+
+        # Training Data
+        if self.config.neuron.local_train:
+            self.dataset = bittensor.dataset(config=self.config)
+            self.dataset.set_data_size(10, 64)
+            data = next(self.dataset)
+
+        # load our old model
+        if not self.config.neuron.restart :
+            self.model.load(self.config.neuron.full_path)
+
+        if self.config.wandb.api_key != 'default':
+            # --- Init Wandb.
+            bittensor.wandb(
+                config = self.config,
+                cold_pubkey = self.wallet.coldkeypub.ss58_address,
+                hot_pubkey = self.wallet.hotkey.ss58_address,
+                root_dir = self.config.neuron.full_path
+            )
+
+        last_set_block = self.subtensor.get_current_block()
+        blocks_per_epoch = self.subtensor.blocks_per_epoch if self.config.neuron.blocks_per_epoch == -1 else self.config.neuron.blocks_per_epoch
+        blocks_per_set_weights = self.subtensor.blocks_per_epoch if self.config.neuron.blocks_per_set_weights == -1 else self.config.neuron.blocks_per_set_weights
+
+        # --- Run Forever.
+        while True:
+            iteration = 0
+            local_data = {}
+            nn = self.subtensor.neuron_for_pubkey(self.wallet.hotkey.ss58_address)
+            uid = self.metagraph.hotkeys.index( self.wallet.hotkey.ss58_address )
+            current_block = self.subtensor.get_current_block()
+            end_block = current_block + self.config.neuron.blocks_per_epoch
+            if self.config.neuron.local_train:
+                # --- Training step.
+                while end_block >= current_block:
+                    if current_block != self.subtensor.get_current_block() and self.axon.priority_threadpool.is_empty:
+                        with self.mutex:
+                            logger.info(f'local training\titeration: {iteration}\tstart')
+                            loss, _ = self.model( next(self.dataset).to(self.model.device) )
+                            if iteration > 0 : 
+                                losses += loss
+                            else:
+                                losses = loss
+                            iteration += 1
+                            current_block = self.subtensor.get_current_block()
+                            logger.info(f'local training\titeration: {iteration}\tloss: {loss}')
+                    else:
+                        time.sleep(1)
+                
+                if iteration != 0:
+                    (losses/iteration).backward()
+            
+            else:
+                while end_block >= current_block:
+                    time.sleep(12)
+                    current_block = self.subtensor.get_current_block()
+
+            # --- Update parameters
+            if (self.config.neuron.local_train and iteration > 0) or (self.config.neuron.remote_train and self.model.backward_gradients_count > 0):
+                # Custom learning rate
+                if self.model.backward_gradients_count > 0:
+                    optimizer.param_groups[0]['lr'] =  0.1/(self.model.backward_gradients_count)
+                else:
+                    optimizer.param_groups[0]['lr'] =  0.1
+
+                logger.info('Optmization Started')
+                with self.mutex:
+                    clip_grad_norm_(self.model.parameters(), 1.0)
+                    optimizer.step()
+                    optimizer.zero_grad()
+                logger.info('Optimization Successful: Model updated')
+
+                if (self.config.neuron.local_train and iteration > 0):
+                    local_data = {'local/loss': losses.detach().item() / iteration}
+
+                    if local_data['local/loss'] < self.model.best_loss:
+                        self.model.best_loss = local_data['local/loss']
+                        self.model.save(self.config.neuron.full_path)
+
+                # Save it only when it gives a low average loss over a large sample size (config.neuron.num_remote_loss), default to 20. 
+                elif (self.config.neuron.remote_train and len(self.model.remote_losses) >= self.config.neuron.num_remote_loss):
+                    local_data = {'local/remote_loss': sum(self.model.remote_losses) / len(self.model.remote_losses)}
+
+                    if local_data['local/remote_loss'] < self.model.best_remote_loss:
+                        self.model.best_remote_loss = local_data['local/remote_loss']
+                        self.model.save(self.config.neuron.full_path)
+
+                    self.model.remote_losses = []
+
+                self.model.backward_gradients_count = 0
+                
+            wandb_data = {            
+                'stake': nn.stake,
+                'rank': nn.rank,
+                'trust': nn.trust,
+                'consensus': nn.consensus,
+                'incentive': nn.incentive,
+                'emission': nn.emission,
+            }
+            
+            if self.config.wandb.api_key != 'default':
+
+                df = pandas.concat( [
+                    bittensor.utils.indexed_values_to_dataframe( prefix = 'w_i_{}'.format(nn.uid), index = self.metagraph.uids, values = self.metagraph.W[:, uid] ),
+                    self.axonaxon.to_dataframe( metagraph = self.metagraph ),
+                ], axis = 1)
+                df['uid'] = df.index
+                wandb_info_axon = self.axon.to_wandb()                
+                wandb.log( { **wandb_data, **wandb_info_axon, **local_data }, step = current_block )
+                wandb.log( { 'stats': wandb.Table( dataframe = df ) }, step = current_block )
+
+            # === Prometheus logging.
+            self.prometheus_guages.labels("stake").set( nn.stake )
+            self.prometheus_guages.labels("rank").set( nn.rank )
+            self.prometheus_guages.labels("trust").set( nn.trust )
+            self.prometheus_guages.labels("consensus").set( nn.consensus )
+            self.prometheus_guages.labels("incentive").set( nn.incentive )
+            self.prometheus_guages.labels("emission").set( nn.emission )
+
+            if current_block - last_set_block > blocks_per_set_weights:
+                bittensor.__console__.print('[green]Current Status:[/green]', {**wandb_data, **local_data})
+                self.metagraph.sync()
+                last_set_block = current_block
+                if not self.config.neuron.no_set_weights:
+                    try: 
+                        bittensor.__console__.print('[green]Current Status:[/green]', {**wandb_data, **local_data})
+                        # Set self weights to maintain activity.
+                        # --- query the chain for the most current number of peers on the network
+                        chain_weights = torch.zeros(self.subtensor.n)
+                        chain_weights [ uid ] = 1 
+                        did_set = self.subtensor.set_weights(
+                            uids=torch.arange(0,self.subtensor.n),
+                            weights = chain_weights,
+                            wait_for_inclusion = False,
+                            wallet = self.wallet,
+                        )
+                        if did_set:
+                            logger.success('Successfully set weights on the chain')
+                        else:
+                            logger.error('Failed to set weights on chain. (Timeout)')
+                        
+                    except Exception as e:
+                        logger.error('Failure setting weights on chain with error: {}', e)
+
+    def synapse_check(self, synapse, hotkey):
+        """
+            Custom synapse function to protect certain synapse functions depending on the stake and weight.
+            Certain synapses require more compute than others. For instance, TEXT_SEQ_2_SEQ requires a significantly
+            more commitment by the server than a requeset for TEXT_CAUSAL_LM_NEXT.
+
+            Args:
+                synapse (:obj:`bittensor.proto.SynapseArgs`, `required`): 
+                    The proto message that contains additional args for individual synapse functions
+                hotkey (:obj:`torch.FloatTensor`, `required`):
+                    The hotkey that sent the request
+
+        """
+        ## Uid that sent the request
+        incoming_uid = self.metagraph.hotkeys.index(hotkey)
+        if synapse.synapse_type == bittensor.proto.Synapse.SynapseType.TEXT_LAST_HIDDEN_STATE:
+            
+            if self.metagraph.S[incoming_uid] < self.config.neuron.lasthidden_stake:
+                return False
+            
+        elif synapse.synapse_type == bittensor.proto.Synapse.SynapseType.TEXT_CAUSAL_LM:
+
+            if self.metagraph.S[incoming_uid] < self.config.neuron.causallm_stake:
+                return False
+
+        elif synapse.synapse_type == bittensor.proto.Synapse.SynapseType.TEXT_CAUSAL_LM_NEXT:
+
+            if self.metagraph.S[incoming_uid] < self.config.neuron.causallmnext_stake:
+                return False
+
+        elif synapse.synapse_type == bittensor.proto.Synapse.SynapseType.TEXT_SEQ_2_SEQ:
+
+            if (self.metagraph.S[incoming_uid] < self.config.neuron.seq2seq_stake) and (self.metagraph.S[incoming_uid,  self.uid]):
+                return False     
+        else:
+            return False
+
+        return True
+
+    def backward_callback(self, inputs_x:torch.FloatTensor, grads_dy:torch.FloatTensor, synapses=[] ):
+        """
+            The default backward callback when no callback is attached: Is used to call specific synapse functions
+
+            Args:
+                inputs_x (:obj:`torch.FloatTensor`, `required`): 
+                    The inputs that will be passed to the synapse functions
+                grads_dy (:obj:`torch.FloatTensor`, `required`):
+                    The gradients that will be passed to the synapse functions
+                synapses (:obj: list of bittensor.proto.SynapseArgs, 'Optional')
+                    The proto message that contains additional args for individual synapse functions
+
+            Returns:
+                response_tensors: (:obj: list of bittensor.proto.Tensor, `required`): 
+                    serialized tensor response from the nucleus call or None.
+                response_codes: (:obj: list of bittensor.proto.ReturnCode, `required`)
+                    return code associated with forward call i.e. Success of Timeout.
+                response_messages: (:obj: list of strings, `required`)
+                    return message associated with synapse call
+        """
+        # --- initialize response variables --- 
+        response_tensors = []
+        response_codes = []
+        response_messages = []
+        
+        if not self.config.neuron.remote_train:
+            return response_tensors, response_codes, response_messages
+
+        # --- calling attached synapses ---
+        with self.mutex and torch.enable_grad() and torch.autograd.set_detect_anomaly(True):
+            for index, synapse in enumerate(synapses):
+                try:
+                    if synapse.synapse_type in self.axon.synapse_callbacks and self.axon.synapse_callbacks[synapse.synapse_type] != None:
+                        message, model_output, response_tensor = self.axon.synapse_callbacks[synapse.synapse_type](inputs_x[index], synapse)
+                        grads_dy_norm = grads_dy[index]/(grads_dy[index].sum() + 0.00001)
+                        torch.autograd.backward (
+                            tensors = [ response_tensor ],
+                            grad_tensors = [ grads_dy_norm ],
+                            retain_graph=True
+                        )
+                        # Only consider loss from causal LM next.
+                        if synapse.synapse_type == bittensor.proto.Synapse.SynapseType.TEXT_CAUSAL_LM_NEXT:
+                            self.model.remote_losses.append(model_output.loss)
+                            self.model.remote_losses = self.model.remote_losses[-self.config.neuron.num_remote_loss:] if len(self.model.remote_losses) > self.config.neuron.num_remote_loss else self.model.remote_losses
+                        self.model.backward_gradients_count += inputs_x[index].size(0)
+                        response_tensors.append(None)
+                        response_codes.append(bittensor.proto.ReturnCode.Success)
+                        response_messages.append('Success')
+                        
+                    else:
+                        response_tensors.append(None)
+                        response_codes.append(bittensor.proto.ReturnCode.NotImplemented)
+                        response_messages.append('Not Implemented')
+                except Exception as e:
+                    # --- Exception Hit in Synapse ---
+                    response_tensors.append(None)
+                    response_codes.append(bittensor.proto.ReturnCode.UnknownException)
+                    response_messages.append(str(e))
+
+        return response_tensors, response_codes, response_messages
+
+
+    def priority(self, pubkey:str, request_type:bittensor.proto.RequestType, inputs_x) -> float:
+        r"""Calculates the priority on requests based on stake and size of input
+            Args:
+                pubkey ( str, `required`):
+                    The public key of the caller.
+                inputs_x ( :obj:`torch.Tensor`, `required`):
+                    torch inputs to be forward processed.
+                request_type ( bittensor.proto.RequestType, `required`):
+                    the request type ('FORWARD' or 'BACKWARD').
+        """
+        try:        
+            uid = self.metagraph.hotkeys.index(pubkey)
+            priority = self.metagraph.S[uid].item()
+        
+        except:
+            # zero priority for those who are not registered.
+            priority =  0
+
+        return priority
+
+    def forward_generate(self, inputs_x:torch.FloatTensor, synapse, model_output = None):
+        tokens = self.model.token_remap(inputs_x.to(self.model.device))
+        output = self.model.pre_model.generate(
+            input_ids=tokens['input_ids'],
+            attention_mask=tokens['attention_mask'],
+            max_length=max(tokens['input_ids'].shape[1] + 1, synapse.num_to_generate),
+            num_beams=synapse.num_beams,
+            no_repeat_ngram_size=synapse.no_repeat_ngram_size,
+            early_stopping = synapse.early_stopping,
+            do_sample=synapse.do_sample,
+            top_p=synapse.top_p,
+            num_return_sequences=synapse.num_return_sequences,
+            temperature = synapse.temperature,
+            repetition_penalty = synapse.repetition_penalty,
+            length_penalty = synapse.length_penalty,
+            max_time = synapse.max_time,
+            num_beam_groups = synapse.num_beam_groups,
+        )
+        raw_texts = [self.model.tokenizer.decode(out) for out in output]
+        tokens = [self.model.std_tokenizer.encode(raw_text, return_tensors="pt")[:,:synapse.num_to_generate].view(-1) for raw_text in raw_texts]
+        bittensor_output = pad_sequence(tokens, batch_first=True)
+        return None, model_output, bittensor_output
+
+    def forward_hidden_state(self, inputs_x:torch.FloatTensor, synapse, model_output = None):
+        with self.mutex:
+            message, model_output, hidden = self.model.encode_forward(inputs_x.to(self.model.device), model_output=model_output)
+        return message, model_output, hidden
+
+    def forward_casual_lm(self, inputs_x:torch.FloatTensor, synapse, model_output = None):
+        with self.mutex:
+            message, model_output, logits = self.model.encode_forward_causallm(inputs_x.to(self.model.device), model_output=model_output)
+        return message, model_output, logits
+
+    def forward_casual_lm_next(self,inputs_x: torch.FloatTensor, synapse, model_output=None):
+        with self.mutex:
+            message, model_output, topk_token_phrases = self.model.encode_forward_causallmnext(inputs_x.to(self.model.device),
+                                                                                        topk=synapse.topk,
+                                                                                        model_output=model_output)
+        # topk_token_phrases: [sum_b(sum_k(len(phrase_k) + 1)_b)] contains topk token phrases and probabilities
+        #   Compacted 1-D tensor >= batch_size * (2 * topk + 1)
+        print('Returned')
+        return message, model_output, topk_token_phrases
+
+
+    def blacklist(self, pubkey:str, request_type:bittensor.proto.RequestType) -> bool:
+        r"""Axon security blacklisting, used to blacklist message from low stake members
+            Args:
+                pubkey ( str, `required`):
+                    The public key of the caller.
+                request_type ( bittensor.proto.RequestType, `required`):
+                    the request type ('FORWARD' or 'BACKWARD').
+        """
+        # Check for registrations
+
+        def registration_check():
+            # If we allow non-registered requests return False = not blacklisted.
+            is_registered = pubkey in self.metagraph.hotkeys
+            if not is_registered:
+                if self.config.neuron.blacklist_allow_non_registered:
+                    return False
+
+                self.prometheus_counters.labels("blacklisted.registration").inc()
+
+                raise Exception('Registration blacklist')
+
+        # Check for stake
+        def stake_check() -> bool:
+            # Check stake.
+            uid = self.metagraph.hotkeys.index(pubkey)
+            if self.metagraph.S[uid].item() < self.config.neuron.blacklist.stake:
+                self.prometheus_counters.labels("blacklisted.stake").inc()
+
+                raise Exception('Stake blacklist')
+            return False
+
+        # Check for time
+        def time_check():
+            current_time = datetime.now()
+            # Only check if the request are forward requests
+            timecheck = self.timecheck_dicts[request_type]
+            if pubkey in timecheck.keys():
+                prev_time = timecheck[pubkey]
+                if current_time - prev_time >= timedelta(seconds=self.config.neuron.blacklist.time):
+                    timecheck[pubkey] = current_time
+                else:
+                    timecheck[pubkey] = current_time
+                    self.prometheus_counters.labels("blacklisted.time").inc()
+
+                    raise Exception('Time blacklist')
+            else:
+                timecheck[pubkey] = current_time
+        
+            return False
+
+        # Black list or not
+        try:
+            registration_check()
+            time_check()
+            stake_check()            
+            return False
+        except Exception as e:
+            self.prometheus_counters.labels("blacklisted").inc()
+            return True
